Pitch slider working
Max pitch scale is 4 but is pretty much the same as 3, so I have the slider maxed out at 3 as it allows for more room and minor adjustments at lower pitch.
Due to how the pitch scale works, it also speeds up the audio file, I don't think I can fix this outside of making different audio buses and having an effect in all of them.